Fix more A11 compilation: trailing commas, cornerRadius Float, TaskStackChangeListener() constructor

+++ app/src/main/java/com/boringdroid/systemui/peek/PeekPanelWindow.kt
@@ -112,7 +112,7 @@ class PeekPanelWindow(
 
         val background = GradientDrawable()
         background.setColor(Color.parseColor("#FF1C1B1F"))
-        background.cornerRadius = dp(8f)
+        background.cornerRadius = dp(8f).toFloat()
         rootLayout.background = background
 
         val layoutParams = LinearLayout.LayoutParams(
